Update Nagios configuration when scaling up the cluster
A cronjob is responsible to scrutinize the /etc/hiera/nodes.yaml to
detect if new node has been added, then re-apply the puppet manifest.
This is the only way we have with MOS 7.0 to add some dynamic
reconfiguration when adding a new node to the cluster.

To notice that this doesn't work when removing a node since the nodes.yaml is
never updated at this stage.
See for details bug https://bugs.launchpad.net/fuel/+bug/1491027

Change-Id: I9eb75a66a55949f2d852fbcd19a4dc67df1dca25