Re-execute necessary tasks on the deployed nodes
This change makes sure that Pacemaker is reconfigured on the nodes that
are already deployed when the environment is scaled up. We use the
'update_required' stanza to enforce that the nodes already deployed are
updated in the same run than the new nodes. The 'reexecute_on' stanza
doesn't guarantee that.

Change-Id: Id8ea87616fa3c9026462a41711732285624d9533
Closes-Bug: #1536533