Prevent duplicated links in dashboard
This changes create the file /var/cache/nagios_link_created for all
nodes. So if you redeploy the environment and another node has the
primary node the links to Grafana won't be pushed again. It is a partial
fix because if you add a new node and if this new node is the primary
node then the link will be pushed and so duplicated in the UI.

Change-Id: I90f641c7c5e6390851dd0a5603406a5b71aecb2a
Partial-Bug: #1544120