Fix deployment when using network templates
Apache can be deployed on any network thanks to network templates. This
change removes all the hard-coded dependencies on the management
network. All port numbers for Apache, Kibana, Elasticsearch, InfluxDB
and Grafana are also moved to Hiera to make it easier to customize if
needed.

It also removes the call to hiera('nodes') since this entry is
deprecated as well as SSH checks because they're not reliable when
network templates are used.

Change-Id: I8f51bda7fee614ac814f0413f8c655b1356b4036
Closes-Bug: #1577358
Closes-Bug: #1550253
Related-Bug: #1514365