Use 2 Virtual IPs for Nagios
One VIP is used for the Nagios UI and the other for the /status
location exclusively used by LMA collectors to report statuses.
This split will help to enable HTTPS for the Nagios UI and deploy it on
another network while LMA collectors continue to use HTTP over the
"monitoring" network.

These VIPs are allocated by default on the management network.

Change-Id: I3c0df769cb40b540b1afb74d4958e87b58daa12d