Configure the management gateway in the network ns
Since Apache runs in its own namespace, it has no default gateway and
can't connect to remote services such as LDAP if those aren't in the
same subnet. As long as one of the 2 VIP addresses is located on the
management network, we should configure the management vrouter as the
default gateway in the infrastructure_alerting network namespace.

Change-Id: I13e5374ad1ca9a6eb17a0600186fe8dd089f2284