Fix a cycle in graph of tasks
This patch removes a dependency that causes a cycle in the graph of
tasks in MOS 8. The cycle was <custom>-netconfig -> hosts -> cluster ->
virtual_ips -> <custom>-netconfig.

We cannot remove hosts.pp as we tried previously because doing this
also remove a dependency between globals and hosts. So we create the
following dependency: tools -> <custom>-netconfig -> hosts
in place of the former: tools -> netconfig -> hosts

Change-Id: I21f6c04eb38406a22a3a400975233bfc775357d0
Partial-Bug: #1616391